[#6] Separated out familia/types
