+++ lib/familia/redisobject.rb
@@ -360,682 +360,11 @@ module Familia
     end
   end
 
-  class List < RedisObject
-    def size
-      redis.llen rediskey
-    end
-    alias length size
+  require_relative 'types/list'
 
-    def empty?
-      size == 0
-    end
+  require_relative 'types/unsorted_set'
+  require_relative 'types/sorted_set'
+  require_relative 'types/hashkey'
+  require_relative 'types/string'
 
-    def push *values
-      echo :push, caller[0] if Familia.debug
-      values.flatten.compact.each { |v| redis.rpush rediskey, to_redis(v) }
-      redis.ltrim rediskey, -@opts[:maxlength], -1 if @opts[:maxlength]
-      update_expiration
-      self
-    end
-
-    def <<(val)
-      push val
-    end
-    alias add <<
-
-    def unshift *values
-      values.flatten.compact.each { |v| redis.lpush rediskey, to_redis(v) }
-      # TODO: test maxlength
-      redis.ltrim rediskey, 0, @opts[:maxlength] - 1 if @opts[:maxlength]
-      update_expiration
-      self
-    end
-
-    def pop
-      from_redis redis.rpop(rediskey)
-    end
-
-    def shift
-      from_redis redis.lpop(rediskey)
-    end
-
-    def [](idx, count = nil)
-      if idx.is_a? Range
-        range idx.first, idx.last
-      elsif count
-        case count <=> 0
-        when 1  then range(idx, idx + count - 1)
-        when 0  then []
-        when -1 then nil
-        end
-      else
-        at idx
-      end
-    end
-    alias slice []
-
-    def delete(v, count = 0)
-      redis.lrem rediskey, count, to_redis(v)
-    end
-    alias remove delete
-    alias rem delete
-    alias del delete
-
-    def range(sidx = 0, eidx = -1)
-      el = rangeraw sidx, eidx
-      multi_from_redis(*el)
-    end
-
-    def rangeraw(sidx = 0, eidx = -1)
-      redis.lrange(rediskey, sidx, eidx)
-    end
-
-    def members(count = -1)
-      echo :members, caller[0] if Familia.debug
-      count -= 1 if count > 0
-      range 0, count
-    end
-    alias all members
-    alias to_a members
-
-    def membersraw(count = -1)
-      count -= 1 if count > 0
-      rangeraw 0, count
-    end
-
-    def each(&blk)
-      range.each(&blk)
-    end
-
-    def each_with_index(&blk)
-      range.each_with_index(&blk)
-    end
-
-    def eachraw(&blk)
-      rangeraw.each(&blk)
-    end
-
-    def eachraw_with_index(&blk)
-      rangeraw.each_with_index(&blk)
-    end
-
-    def collect(&blk)
-      range.collect(&blk)
-    end
-
-    def select(&blk)
-      range.select(&blk)
-    end
-
-    def collectraw(&blk)
-      rangeraw.collect(&blk)
-    end
-
-    def selectraw(&blk)
-      rangeraw.select(&blk)
-    end
-
-    def at(idx)
-      from_redis redis.lindex(rediskey, idx)
-    end
-
-    def first
-      at 0
-    end
-
-    def last
-      at(-1)
-    end
-
-    # TODO: def replace
-    ## Make the value stored at KEY identical to the given list
-    # define_method :"#{name}_sync" do |*latest|
-    #  latest = latest.flatten.compact
-    #  # Do nothing if we're given an empty Array.
-    #  # Otherwise this would clear all current values
-    #  if latest.empty?
-    #    false
-    #  else
-    #    # Convert to a list of index values if we got the actual objects
-    #    latest = latest.collect { |obj| obj.index } if klass === latest.first
-    #    current = send("#{name_plural}raw")
-    #    added = latest-current
-    #    removed = current-latest
-    #    #Familia.info "#{self.index}: adding: #{added}"
-    #    added.each { |v| self.send("add_#{name_singular}", v) }
-    #    #Familia.info "#{self.index}: removing: #{removed}"
-    #    removed.each { |v| self.send("remove_#{name_singular}", v) }
-    #    true
-    #  end
-    # end
-
-    Familia::RedisObject.register self, :list
-  end
-
-  class Set < RedisObject
-    def size
-      redis.scard rediskey
-    end
-    alias length size
-
-    def empty?
-      size == 0
-    end
-
-    def add *values
-      values.flatten.compact.each { |v| redis.sadd? rediskey, to_redis(v) }
-      update_expiration
-      self
-    end
-
-    def <<(v)
-      add v
-    end
-
-    def members
-      echo :members, caller[0] if Familia.debug
-      el = membersraw
-      multi_from_redis(*el)
-    end
-    alias all members
-    alias to_a members
-
-    def membersraw
-      redis.smembers(rediskey)
-    end
-
-    def each(&blk)
-      members.each(&blk)
-    end
-
-    def each_with_index(&blk)
-      members.each_with_index(&blk)
-    end
-
-    def collect(&blk)
-      members.collect(&blk)
-    end
-
-    def select(&blk)
-      members.select(&blk)
-    end
-
-    def eachraw(&blk)
-      membersraw.each(&blk)
-    end
-
-    def eachraw_with_index(&blk)
-      membersraw.each_with_index(&blk)
-    end
-
-    def collectraw(&blk)
-      membersraw.collect(&blk)
-    end
-
-    def selectraw(&blk)
-      membersraw.select(&blk)
-    end
-
-    def member?(val)
-      redis.sismember rediskey, to_redis(val)
-    end
-    alias include? member?
-
-    def delete(val)
-      redis.srem rediskey, to_redis(val)
-    end
-    alias remove delete
-    alias rem delete
-    alias del delete
-
-    def intersection *setkeys
-      # TODO
-    end
-
-    def pop
-      redis.spop rediskey
-    end
-
-    def move(dstkey, val)
-      redis.smove rediskey, dstkey, val
-    end
-
-    def random
-      from_redis randomraw
-    end
-
-    def randomraw
-      redis.srandmember(rediskey)
-    end
-
-    ## Make the value stored at KEY identical to the given list
-    # define_method :"#{name}_sync" do |*latest|
-    #  latest = latest.flatten.compact
-    #  # Do nothing if we're given an empty Array.
-    #  # Otherwise this would clear all current values
-    #  if latest.empty?
-    #    false
-    #  else
-    #    # Convert to a list of index values if we got the actual objects
-    #    latest = latest.collect { |obj| obj.index } if klass === latest.first
-    #    current = send("#{name_plural}raw")
-    #    added = latest-current
-    #    removed = current-latest
-    #    #Familia.info "#{self.index}: adding: #{added}"
-    #    added.each { |v| self.send("add_#{name_singular}", v) }
-    #    #Familia.info "#{self.index}: removing: #{removed}"
-    #    removed.each { |v| self.send("remove_#{name_singular}", v) }
-    #    true
-    #  end
-    # end
-
-    Familia::RedisObject.register self, :set
-  end
-
-  class SortedSet < RedisObject
-    def size
-      redis.zcard rediskey
-    end
-    alias length size
-
-    def empty?
-      size.zero?
-    end
-
-    # NOTE: The argument order is the reverse of #add
-    # e.g. obj.metrics[VALUE] = SCORE
-    def []=(val, score)
-      add score, val
-    end
-
-    # NOTE: The argument order is the reverse of #[]=
-    def add(score, val)
-      ret = redis.zadd rediskey, score, to_redis(val)
-      update_expiration
-      ret
-    end
-
-    def score(val)
-      ret = redis.zscore rediskey, to_redis(val)
-      ret&.to_f
-    end
-    alias [] score
-
-    def member?(val)
-      !rank(val).nil?
-    end
-    alias include? member?
-
-    # rank of member +v+ when ordered lowest to highest (starts at 0)
-    def rank(v)
-      ret = redis.zrank rediskey, to_redis(v)
-      ret&.to_i
-    end
-
-    # rank of member +v+ when ordered highest to lowest (starts at 0)
-    def revrank(v)
-      ret = redis.zrevrank rediskey, to_redis(v)
-      ret&.to_i
-    end
-
-    def members(count = -1, opts = {})
-      count -= 1 if count.positive?
-      el = membersraw count, opts
-      multi_from_redis(*el)
-    end
-    alias to_a members
-    alias all members
-
-    def membersraw(count = -1, opts = {})
-      count -= 1 if count.positive?
-      rangeraw 0, count, opts
-    end
-
-    def revmembers(count = -1, opts = {})
-      count -= 1 if count.positive?
-      el = revmembersraw count, opts
-      multi_from_redis(*el)
-    end
-
-    def revmembersraw(count = -1, opts = {})
-      count -= 1 if count.positive?
-      revrangeraw 0, count, opts
-    end
-
-    def each(&blk)
-      members.each(&blk)
-    end
-
-    def each_with_index(&blk)
-      members.each_with_index(&blk)
-    end
-
-    def collect(&blk)
-      members.collect(&blk)
-    end
-
-    def select(&blk)
-      members.select(&blk)
-    end
-
-    def eachraw(&blk)
-      membersraw.each(&blk)
-    end
-
-    def eachraw_with_index(&blk)
-      membersraw.each_with_index(&blk)
-    end
-
-    def collectraw(&blk)
-      membersraw.collect(&blk)
-    end
-
-    def selectraw(&blk)
-      membersraw.select(&blk)
-    end
-
-    def range(sidx, eidx, opts = {})
-      echo :range, caller[0] if Familia.debug
-      el = rangeraw(sidx, eidx, opts)
-      multi_from_redis(*el)
-    end
-
-    def rangeraw(sidx, eidx, opts = {})
-      # NOTE: :withscores (no underscore) is the correct naming for the
-      # redis-4.x gem. We pass :withscores through explicitly b/c
-      # redis.zrange et al only accept that one optional argument.
-      # Passing `opts`` through leads to an ArgumentError:
-      #
-      #   sorted_sets.rb:374:in `zrevrange': wrong number of arguments (given 4, expected 3) (ArgumentError)
-      #
-      redis.zrange(rediskey, sidx, eidx, **opts)
-    end
-
-    def revrange(sidx, eidx, opts = {})
-      echo :revrange, caller[0] if Familia.debug
-      el = revrangeraw(sidx, eidx, opts)
-      multi_from_redis(*el)
-    end
-
-    def revrangeraw(sidx, eidx, opts = {})
-      redis.zrevrange(rediskey, sidx, eidx, **opts)
-    end
-
-    # e.g. obj.metrics.rangebyscore (now-12.hours), now, :limit => [0, 10]
-    def rangebyscore(sscore, escore, opts = {})
-      echo :rangebyscore, caller[0] if Familia.debug
-      el = rangebyscoreraw(sscore, escore, opts)
-      multi_from_redis(*el)
-    end
-
-    def rangebyscoreraw(sscore, escore, opts = {})
-      echo :rangebyscoreraw, caller[0] if Familia.debug
-      redis.zrangebyscore(rediskey, sscore, escore, **opts)
-    end
-
-    # e.g. obj.metrics.revrangebyscore (now-12.hours), now, :limit => [0, 10]
-    def revrangebyscore(sscore, escore, opts={})
-      echo :revrangebyscore, caller[0] if Familia.debug
-      el = revrangebyscoreraw(sscore, escore, opts)
-      multi_from_redis(*el)
-    end
-
-    def revrangebyscoreraw(sscore, escore, opts={})
-      echo :revrangebyscoreraw, caller[0] if Familia.debug
-      opts[:with_scores] = true if opts[:withscores]
-      redis.zrevrangebyscore(rediskey, sscore, escore, opts)
-    end
-
-    def remrangebyrank(srank, erank)
-      redis.zremrangebyrank rediskey, srank, erank
-    end
-
-    def remrangebyscore(sscore, escore)
-      redis.zremrangebyscore rediskey, sscore, escore
-    end
-
-    def increment(val, by = 1)
-      redis.zincrby(rediskey, by, val).to_i
-    end
-    alias incr increment
-    alias incrby increment
-
-    def decrement(val, by = 1)
-      increment val, -by
-    end
-    alias decr decrement
-    alias decrby decrement
-
-    def delete(val)
-      redis.zrem rediskey, to_redis(val)
-    end
-    alias remove delete
-    alias rem delete
-    alias del delete
-
-    def at(idx)
-      range(idx, idx).first
-    end
-
-    # Return the first element in the list. Redis: ZRANGE(0)
-    def first
-      at(0)
-    end
-
-    # Return the last element in the list. Redis: ZRANGE(-1)
-    def last
-      at(-1)
-    end
-
-    Familia::RedisObject.register self, :zset
-  end
-
-  class HashKey < RedisObject
-    def size
-      redis.hlen rediskey
-    end
-    alias length size
-
-    def empty?
-      size == 0
-    end
-
-    def []=(name, val)
-      ret = redis.hset rediskey, name, to_redis(val)
-      update_expiration
-      ret
-    rescue TypeError => e
-      echo :hset, caller[0] if Familia.debug
-      klass = val.class
-      msg = "Cannot store #{name} => #{val.inspect} (#{klass}) in #{rediskey}"
-      raise e.class, msg
-    end
-    alias put []=
-    alias store []=
-
-    def [](name)
-      from_redis redis.hget(rediskey, name)
-    end
-    alias get []
-
-    def fetch(name, default = nil)
-      ret = self[name]
-      if ret.nil?
-        raise IndexError, "No such index for: #{name}" if default.nil?
-
-        default
-      else
-        ret
-      end
-    end
-
-    def keys
-      redis.hkeys rediskey
-    end
-
-    def values
-      el = redis.hvals(rediskey)
-      multi_from_redis(*el)
-    end
-
-    def all
-      # TODO: Use from_redis. Also name `all` is confusing with
-      # Onetime::Customer.all which returns all customers.
-      redis.hgetall rediskey
-    end
-    alias to_hash all
-    alias clone all
-
-    def has_key?(name)
-      redis.hexists rediskey, name
-    end
-    alias include? has_key?
-    alias member? has_key?
-
-    def delete(name)
-      redis.hdel rediskey, name
-    end
-    alias remove delete
-    alias rem delete
-    alias del delete
-
-    def increment(name, by = 1)
-      redis.hincrby(rediskey, name, by).to_i
-    end
-    alias incr increment
-    alias incrby increment
-
-    def decrement(name, by = 1)
-      increment name, -by
-    end
-    alias decr decrement
-    alias decrby decrement
-
-    def update(hsh = {})
-      raise ArgumentError, 'Argument to bulk_set must be a hash' unless hsh.is_a?(Hash)
-
-      data = hsh.inject([]) { |ret, pair| ret << [pair[0].class, to_redis(pair[1]).class] }.flatten
-      ret = redis.hmset(rediskey, *data)
-      update_expiration
-      ret
-    end
-    alias merge! update
-
-    def values_at *names
-      el = redis.hmget(rediskey, *names.flatten.compact)
-      multi_from_redis(*el)
-    end
-
-    Familia::RedisObject.register self, :hash
-  end
-
-  class String < RedisObject
-    def init; end
-
-    def size
-      to_s.size
-    end
-    alias length size
-
-    def empty?
-      size.zero?
-    end
-
-    def value
-      echo :value, caller[0..5] if Familia.debug
-      redis.setnx rediskey, @opts[:default] if @opts[:default]
-      from_redis redis.get(rediskey)
-    end
-    alias content value
-    alias get value
-
-    def to_s
-      value.to_s # value can return nil which to_s should not
-    end
-
-    def to_i
-      value.to_i
-    end
-
-    def value=(val)
-      ret = redis.set(rediskey, to_redis(val))
-      update_expiration
-      ret
-    end
-    alias replace value=
-    alias set value=
-
-    def setnx(val)
-      ret = redis.setnx(rediskey, to_redis(val))
-      update_expiration
-      ret
-    end
-
-    def increment
-      ret = redis.incr(rediskey)
-      update_expiration
-      ret
-    end
-    alias incr increment
-
-    def incrementby(val)
-      ret = redis.incrby(rediskey, val.to_i)
-      update_expiration
-      ret
-    end
-    alias incrby incrementby
-
-    def decrement
-      ret = redis.decr rediskey
-      update_expiration
-      ret
-    end
-    alias decr decrement
-
-    def decrementby(val)
-      ret = redis.decrby rediskey, val.to_i
-      update_expiration
-      ret
-    end
-    alias decrby decrementby
-
-    def append(val)
-      ret = redis.append rediskey, val
-      update_expiration
-      ret
-    end
-    alias << append
-
-    def getbit(offset)
-      redis.getbit rediskey, offset
-    end
-
-    def setbit(offset, val)
-      ret = redis.setbit rediskey, offset, val
-      update_expiration
-      ret
-    end
-
-    def getrange(spoint, epoint)
-      redis.getrange rediskey, spoint, epoint
-    end
-
-    def setrange(offset, val)
-      ret = redis.setrange rediskey, offset, val
-      update_expiration
-      ret
-    end
-
-    def getset(val)
-      ret = redis.getset rediskey, val
-      update_expiration
-      ret
-    end
-
-    def nil?
-      value.nil?
-    end
-
-    Familia::RedisObject.register self, :string
-  end
 end
